create competition component and some other things
