+++ src/react/components/Footer.css
@@ -4,7 +4,7 @@
   justify-content: center;
 }
 
-.footer-text {
+.footer__text {
   margin: 1vh 2vh;
   color: black;
   font-size: large;
